I THINK it works?
one of the problems was the fact that that the matrices being entered into teh sort function were different sizes. This wasn't a problem for Nasim because the number of components was limited by the Y matrix, which I don't have.

I "fixed" it by ignoring the zero index. Not sure if that actually solves the problem, though.